core, fix: trigger keyword for use-package
Some of our triggers are pre-command-hook. I noticed when we open Emacs,
run any command, there'll be an error says the hook variable in function
toki|after-call* is void. It looks like that function (created by
toki/add-trigger) didn't capture the argument HOOK-OR-FUNCTION, but
actually it works well regardless of the error. Anyway, here's a fix. We
expand everything in the use-package form, and offers the value of
HOOK-OR-FUNCTION explicitely to toki/add-trigger.