tabs, enh: don't update in buffer-list-update-hook
This could cause performance issue when a lot of buffer is created.
Sometimes when I visit a file using TRAMP, I have a call chain like

toki-tabs/update-buffer-list -> toki-tabs/buffer-group ->
project-current -> project-try-vc -> generate-new-buffer ->
buffer-list-update-hook -> toki-tabs/update-buffer-list

Which resuls in an infinite loop.

The new implementation only updates in post-command-hook and in an idle
timer, and when certain conditions are satisfied so it should fix the
performance issue.